Refactor agent gateway client
