--- internal/client/rest/agent/option.go
@@ -1,17 +1,0 @@
-package agent
-
-type Option func(*agentClient)
-
-var (
-	defaultOptions = []Option{
-		WithAddr("0.0.0.0:8081"),
-	}
-)
-
-func WithAddr(addr string) Option {
-	return func(ac *agentClient) {
-		if len(addr) != 0 {
-			ac.addr = addr
-		}
-	}
-}
